refactor: refactor settings into config store
