+++ src/container.ts
@@ -8,6 +8,7 @@ import EpgService from '#src/services/epg.service';
 import EntitlementService from '#src/services/entitlement.service';
 import FavoritesService from '#src/services/favorites.service';
 import ConfigService from '#src/services/config.service';
+import SettingsService from '#src/services/settings.service';
 
 
 import ApiController from '#src/stores/ApiController';
@@ -19,7 +20,6 @@ import EntitlementController from '#src/stores/EntitlementController';
 import ProfileController from '#src/stores/ProfileController';
 import FavoritesController from '#src/stores/FavoritesController';
 import AppController from '#src/stores/AppController';
-import { SettingsController } from '#src/stores/SettingsController';
 import type { interfaces } from 'inversify';
 
 export const container = new Container({ defaultScope: 'Singleton', skipBaseClassChecks: true });
@@ -36,16 +36,17 @@ container.bind(WatchHistoryService).toSelf();
 container.bind(FavoritesService).toSelf();
 container.bind(EntitlementService).toSelf();
 container.bind(ApiService).toSelf();
+container.bind(SettingsService).toSelf();
 
 // Common controllers
+container.bind(AppController).toSelf();
 container.bind(ApiController).toSelf();
 container.bind(EpgController).toSelf();
 container.bind(WatchHistoryController).toSelf();
 container.bind(FavoritesController).toSelf();
 container.bind(EntitlementController).toSelf();
 
+// Integration controllers (conditionally register?)
 container.bind(AccountController).toSelf();
 container.bind(CheckoutController).toSelf();
 container.bind(ProfileController).toSelf();
-container.bind(AppController).toSelf();
-container.bind(SettingsController).toSelf();
